update exam
Items 7 to 11 resolved
